Refactor tuner: extract stage runners and decompose hf_jobs_backend
Split experiment_handler.py (946→207 lines) by extracting
HFTrainingStageRunner, HFEvalStageRunner, and HFLossStageRunner
into tuner/handlers/stages/ subpackage. ExperimentHandler remains
as the orchestrator.

Decompose HFJobsBackend (990→465 lines) into 4 focused mixins:
CommandBuilder, JobWatcher, BucketOps, PostTraining. External API
(ITrainingBackend) unchanged.
