+++ tuner/backends/training/cloud/hf_jobs_backend.py
@@ -10,6 +10,10 @@ submitted via huggingface_hub.run_job(image, command, flavor) which
 provisions GPU hardware and runs a Docker container with the specified
 command. The command clones the repo and runs the training script.
 
+This module serves as a facade: HFJobsBackend composes four mixins that
+encapsulate distinct concerns (command building, job watching, bucket
+operations, post-training actions). The external API is unchanged.
+
 Requirements:
 - HF_TOKEN environment variable set (requires HF Pro subscription)
 - huggingface_hub >= 0.27.0 installed
@@ -17,35 +21,21 @@ Requirements:
 
 import logging
 import os
-import shlex
-import shutil
-import sys
-import tempfile
-import time
 import yaml
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import List, Optional, Tuple
 
 from shared.cloud_artifacts import normalize_hf_bucket_id
-from shared.utilities.env import get_hf_token
-from shared.utilities.unique_ids import unique_utc_timestamp
-from shared.utilities.paths import (
-    get_canonical_trainer_dir_name,
-    get_primary_training_output_dir,
-    get_trainer_root,
-)
+from shared.utilities.paths import get_trainer_root
 from tuner.cloud import (
     CloudJobSpec,
     HFJobExecutor,
-    RepoCheckoutSpec,
     build_bash_command,
     build_hf_job_secrets,
-    build_repo_checkout_steps,
     load_huggingface_hub,
     resolve_hf_bucket_id,
 )
-from tuner.ui import BOX, RICH_AVAILABLE, print_config, print_info, print_menu, print_success
+from tuner.ui import print_config
 from tuner.backends.training.base import ITrainingBackend
 from tuner.core.config import TrainingConfig, CloudTrainingConfig
 from tuner.core.exceptions import CloudProviderError, ConfigurationError
@@ -53,11 +43,14 @@ from tuner.core.interfaces import ExecuteResult
 
 from .base_cloud import (
     load_cloud_config,
-    load_project_deps,
     poll_until_done,
     resolve_cloud_image,
     resolve_repo_source,
 )
+from ._hf_command_builder import HFCommandBuilderMixin
+from ._hf_job_watcher import HFJobWatcherMixin
+from ._hf_bucket_ops import HFBucketOpsMixin
+from ._hf_post_training import HFPostTrainingMixin
 
 logger = logging.getLogger(__name__)
 
@@ -67,7 +60,13 @@ DEFAULT_TIMEOUT = "4h"
 DEFAULT_IMAGE = "unsloth/unsloth:2026.1.2-pt2.9.0-cu12.8-update@sha256:5266c57be21059bfb407d80dc2f448868a5c2e2dbe7b2aa27780f48b48cbec39"
 
 
-class HFJobsBackend(ITrainingBackend):
+class HFJobsBackend(
+    HFCommandBuilderMixin,
+    HFJobWatcherMixin,
+    HFBucketOpsMixin,
+    HFPostTrainingMixin,
+    ITrainingBackend,
+):
     """
     HuggingFace Jobs training backend.
 
@@ -100,42 +99,10 @@ class HFJobsBackend(ITrainingBackend):
         self.last_bucket_id: Optional[str] = None
         self.last_job_id: Optional[str] = None
 
-    ENV_GRPO_CONFIG_NAME = "env_config.yaml"
-    ENV_GRPO_SCRIPT_NAME = "train_env_grpo.py"
-
-    @classmethod
-    def _config_filename_for_method(cls, method: str) -> str:
-        return cls.ENV_GRPO_CONFIG_NAME if method == "grpo" else "config.yaml"
-
-    @classmethod
-    def _script_name_for_method(cls, method: str) -> str:
-        return cls.ENV_GRPO_SCRIPT_NAME if method == "grpo" else f"train_{method}.py"
-
-    @staticmethod
-    def _dataset_display(dataset_config: dict) -> str:
-        local_file = dataset_config.get("local_file")
-        if local_file:
-            return str(local_file)
-
-        dataset_name = str(dataset_config.get("dataset_name") or "").strip()
-        dataset_file = str(dataset_config.get("dataset_file") or "").strip()
-        if dataset_name and dataset_file:
-            return f"{dataset_name}/{dataset_file}"
-        if dataset_file:
-            return dataset_file
-        if dataset_name:
-            return dataset_name
-        return "Unknown"
-
     @property
     def name(self) -> str:
         """Backend identifier."""
         return "hf_jobs"
-
-    @staticmethod
-    def _new_run_timestamp() -> str:
-        """Return a launch timestamp with a short nonce to avoid same-second collisions."""
-        return unique_utc_timestamp()
 
     def get_available_methods(self) -> List[str]:
         """
@@ -297,6 +264,7 @@ class HFJobsBackend(ITrainingBackend):
             artifact_backend=hf_config.get("artifact_backend", "hf_bucket"),
             artifact_identifier=hf_config.get("artifact_identifier"),
             artifact_mount_path=hf_config.get("output_root", "/workspace/outputs"),
+            pip_packages=hf_config.get("pip_packages"),
             publish_final_model=artifacts_config.get("publish_final_model", False),
             publish_target_repo=artifacts_config.get("publish_target_repo"),
             repo_url=repo_source.url,
@@ -457,499 +425,6 @@ class HFJobsBackend(ITrainingBackend):
         error_detail = result.replace("ERROR: ", "") if result.startswith("ERROR: ") else result
         print(f"  Job {job_id} failed: {error_detail}")
         return _build_result(1)
-
-    def _should_use_remote_dashboard(self, config: CloudTrainingConfig) -> bool:
-        """Only use the live remote dashboard for interactive HF bucket runs."""
-        return (
-            config.artifact_backend == "hf_bucket"
-            and RICH_AVAILABLE
-            and sys.stdin.isatty()
-            and sys.stdout.isatty()
-        )
-
-    def _watch_job_with_remote_dashboard(
-        self,
-        *,
-        config: CloudTrainingConfig,
-        huggingface_hub,
-        job_id: str,
-        artifact_prefix: str,
-    ) -> int:
-        """Render the existing training dashboard using metrics mirrored to the HF bucket."""
-        from shared.ui import LiveDashboard
-
-        hf_token = get_hf_token()
-        from huggingface_hub import sync_bucket
-
-        timeout_seconds = int(config.timeout_hours * 3600)
-        poll_interval = 15
-        elapsed = 0
-        last_status = None
-        last_job_log_offset = 0
-        processed_remote_lines = 0
-        local_logs_dir = Path(tempfile.mkdtemp(prefix="hf-job-logs-"))
-        dashboard = LiveDashboard(
-            title=f"HF Jobs {config.method.upper()}",
-            training_type=config.method,
-            log_lines=5,
-        )
-
-        try:
-            with dashboard:
-                while elapsed < timeout_seconds:
-                    try:
-                        job_info = huggingface_hub.inspect_job(job_id=job_id)
-                        status_obj = getattr(job_info, "status", None)
-                        status = status_obj.stage if status_obj and hasattr(status_obj, "stage") else str(status_obj or "UNKNOWN")
-                    except Exception as e:
-                        dashboard.update(log_message=f"Status check failed: {e}")
-                        status = last_status or "UNKNOWN"
-
-                    if status != last_status:
-                        dashboard.update(log_message=f"Job status: {status}")
-                        last_status = status
-
-                    try:
-                        logs = huggingface_hub.fetch_job_logs(job_id=job_id) or ""
-                        if len(logs) > last_job_log_offset:
-                            new_lines = [line for line in logs[last_job_log_offset:].splitlines() if line.strip()]
-                            for line in new_lines[-2:]:
-                                dashboard.update(log_message=line)
-                            last_job_log_offset = len(logs)
-                    except Exception:
-                        pass
-
-                    try:
-                        sync_bucket(
-                            f"hf://buckets/{config.artifact_identifier}/{artifact_prefix}/logs",
-                            str(local_logs_dir),
-                            token=hf_token,
-                        )
-                        processed_remote_lines = self._update_dashboard_from_local_log(
-                            local_logs_dir=local_logs_dir,
-                            dashboard=dashboard,
-                            processed_remote_lines=processed_remote_lines,
-                        )
-                    except Exception as e:
-                        dashboard.update(log_message=f"Remote log sync unavailable: {e}")
-
-                    if status in ("completed", "COMPLETED"):
-                        dashboard.update(log_message="Training completed successfully.")
-                        return self._finalize_completed_job(config=config, artifact_prefix=artifact_prefix)
-                    if status in ("error", "ERROR", "failed", "FAILED"):
-                        dashboard.update(log_message=f"Job failed with status: {status}")
-                        return 1
-                    if status in ("cancelled", "CANCELLED"):
-                        dashboard.update(log_message="Job was cancelled.")
-                        return 1
-
-                    time.sleep(poll_interval)
-                    elapsed += poll_interval
-        finally:
-            shutil.rmtree(local_logs_dir, ignore_errors=True)
-
-        recovered = self._recover_completed_run_from_bucket(
-            config=config,
-            artifact_prefix=artifact_prefix,
-        )
-        if recovered:
-            print()
-            print(f"  Job {job_id} appears complete based on synced artifacts.")
-            return self._finalize_completed_job(config=config, artifact_prefix=artifact_prefix)
-
-        print()
-        print(f"  Job {job_id} failed: Timeout exceeded")
-        return 1
-
-    def _update_dashboard_from_local_log(self, *, local_logs_dir: Path, dashboard, processed_remote_lines: int) -> int:
-        """Read mirrored bucket logs from a local temp directory and feed them into the dashboard."""
-        candidates = sorted(local_logs_dir.glob("training_*.jsonl"))
-        if not candidates:
-            return processed_remote_lines
-
-        with candidates[-1].open("r", encoding="utf-8") as handle:
-            lines = [line.strip() for line in handle.readlines() if line.strip()]
-
-        for line in lines[processed_remote_lines:]:
-            dashboard._process_log_line(line)
-
-        return len(lines)
-
-    def _ensure_hf_bucket(self, config: CloudTrainingConfig, huggingface_hub) -> None:
-        """Resolve and create the configured HF bucket before launching the job."""
-        if not config.artifact_identifier:
-            raise CloudProviderError("HF Jobs requires an artifact bucket identifier.")
-
-        hf_token = get_hf_token()
-        if not hf_token:
-            raise CloudProviderError(
-                "HF_TOKEN not set. Required for HuggingFace Jobs bucket creation."
-            )
-        config.artifact_identifier = resolve_hf_bucket_id(
-            huggingface_hub,
-            config.artifact_identifier,
-            token=hf_token,
-            private=True,
-        )
-        logger.info("Using HF bucket: %s", config.artifact_identifier)
-
-    def _build_remote_run_uri(self, config: CloudTrainingConfig, artifact_prefix: str) -> str:
-        """Return the HF bucket URI for a completed run."""
-        if not config.artifact_identifier:
-            raise CloudProviderError("HF Jobs requires an artifact bucket identifier.")
-        return f"hf://buckets/{config.artifact_identifier}/{artifact_prefix.strip('/')}"
-
-    def _local_download_run_dir(self, config: CloudTrainingConfig, artifact_prefix: str) -> Path:
-        """Return the local run directory used for downloaded cloud runs."""
-        run_slug = artifact_prefix.strip("/").split("/")[-1]
-        return get_primary_training_output_dir(config.method, self.repo_root) / run_slug
-
-    def _sync_bucket_path(self, remote_uri: str, local_dir: Path, *, token: Optional[str] = None) -> None:
-        """Sync a remote HF bucket path into a local directory."""
-        from huggingface_hub import sync_bucket
-
-        local_dir = Path(local_dir)
-        local_dir.mkdir(parents=True, exist_ok=True)
-        sync_bucket(remote_uri, str(local_dir), token=token)
-
-    def _run_dir_has_completion_artifacts(self, run_dir: Path) -> bool:
-        """Return True when a run directory appears complete enough for next-step workflows."""
-        final_model_dir = run_dir / "final_model"
-        has_final_model = final_model_dir.exists() and any(final_model_dir.iterdir())
-        has_lineage = (run_dir / "training_lineage.json").exists()
-        has_grpo_logs = (run_dir / "logs" / "training_latest.jsonl").exists()
-        return has_final_model and (has_lineage or has_grpo_logs)
-
-    def _download_completed_run(
-        self,
-        *,
-        config: CloudTrainingConfig,
-        artifact_prefix: str,
-        local_run_dir: Optional[Path] = None,
-    ) -> Path:
-        """Download a completed HF Jobs run into the local training outputs tree."""
-        target_dir = Path(local_run_dir) if local_run_dir else self._local_download_run_dir(config, artifact_prefix)
-        self._sync_bucket_path(
-            self._build_remote_run_uri(config, artifact_prefix),
-            target_dir,
-            token=get_hf_token(),
-        )
-        return target_dir
-
-    def _recover_completed_run_from_bucket(
-        self,
-        *,
-        config: CloudTrainingConfig,
-        artifact_prefix: str,
-    ) -> bool:
-        """Best-effort completion recovery when HF status lags behind uploaded artifacts."""
-        recovery_dir = Path(tempfile.mkdtemp(prefix="hf-job-recovery-"))
-        try:
-            self._sync_bucket_path(
-                self._build_remote_run_uri(config, artifact_prefix),
-                recovery_dir,
-                token=get_hf_token(),
-            )
-            return self._run_dir_has_completion_artifacts(recovery_dir)
-        except Exception as exc:
-            logger.warning("HF Jobs completion recovery failed: %s", exc)
-            return False
-        finally:
-            shutil.rmtree(recovery_dir, ignore_errors=True)
-
-    def _print_completion_summary(
-        self,
-        *,
-        config: CloudTrainingConfig,
-        artifact_prefix: str,
-        local_run_dir: Optional[Path] = None,
-    ) -> None:
-        """Print where the completed artifacts live and how they map locally."""
-        summary = {
-            "Remote artifacts": self._build_remote_run_uri(config, artifact_prefix),
-            "Suggested local path": str(self._local_download_run_dir(config, artifact_prefix)),
-        }
-        if local_run_dir:
-            summary["Local run"] = str(local_run_dir)
-        print_config(summary, "Cloud Training Artifacts")
-
-    def _handle_post_training_actions(self, *, config: CloudTrainingConfig, artifact_prefix: str) -> None:
-        """Offer a post-training menu that reuses the existing local workflows."""
-        if not (sys.stdin.isatty() and sys.stdout.isatty()):
-            return
-
-        from tuner.handlers.convert_handler import ConvertHandler
-        from tuner.handlers.eval_handler import EvalHandler
-        from tuner.handlers.inference_handler import InferenceHandler
-        from tuner.handlers.upload_handler import UploadHandler
-
-        local_run_dir: Optional[Path] = None
-        menu_options = [
-            ("download", f"{BOX['star']} Download completed run locally"),
-            ("gguf", f"{BOX['bullet']} Convert model to GGUF"),
-            ("run", f"{BOX['bullet']} Run model locally"),
-            ("eval", f"{BOX['bullet']} Run evaluation"),
-            ("upload", f"{BOX['bullet']} Upload model to Hugging Face"),
-        ]
-
-        while True:
-            choice = print_menu(menu_options, "Choose your next step:")
-            if not choice:
-                return
-
-            if local_run_dir is None:
-                print_info("Downloading completed cloud run into local training outputs...")
-                local_run_dir = self._download_completed_run(
-                    config=config,
-                    artifact_prefix=artifact_prefix,
-                )
-                print_success(f"Downloaded run to: {local_run_dir}")
-
-            if choice == "download":
-                continue
-            if choice == "gguf":
-                print_info("Opening conversion workflow. Select the newly downloaded run when prompted.")
-                ConvertHandler().handle()
-                continue
-            if choice == "run":
-                print_info("Opening local inference workflow. Select the newly downloaded run when prompted.")
-                InferenceHandler().handle()
-                continue
-            if choice == "eval":
-                print_info("Opening evaluation workflow. Select the newly downloaded run when prompted.")
-                EvalHandler(args=None).handle()
-                continue
-            if choice == "upload":
-                print_info("Opening upload workflow. Select the newly downloaded run when prompted.")
-                UploadHandler(args=None).handle()
-
-    def _finalize_completed_job(
-        self,
-        *,
-        config: CloudTrainingConfig,
-        artifact_prefix: str,
-        local_run_dir: Optional[Path] = None,
-    ) -> int:
-        """Print success details and offer post-training actions."""
-        self._print_completion_summary(
-            config=config,
-            artifact_prefix=artifact_prefix,
-            local_run_dir=local_run_dir,
-        )
-        if self.show_post_training_actions:
-            self._handle_post_training_actions(config=config, artifact_prefix=artifact_prefix)
-        return 0
-
-    def _build_artifact_prefix(self, config: CloudTrainingConfig, timestamp: str) -> str:
-        """Build the canonical artifact prefix used by HF Jobs runs."""
-        if not config.repo_commit:
-            raise CloudProviderError("HF Jobs requires exact repo source metadata.")
-        run_slug = f"{timestamp}-{config.repo_commit[:8]}"
-        return f"runs/{config.provider}/{config.method}/{run_slug}"
-
-    def _build_training_command(self, config: CloudTrainingConfig, timestamp: str | None = None) -> str:
-        """
-        Build a shell command string for the HF Jobs container.
-
-        The command installs training dependencies, clones the repo,
-        and runs the training script. HF auth is passed via job secrets
-        when the job is submitted.
-
-        Args:
-            config: Cloud training configuration
-
-        Returns:
-            Shell command string to pass as ["bash", "-c", command]
-        """
-        if not config.repo_url or not config.repo_branch or not config.repo_commit:
-            raise CloudProviderError("HF Jobs requires exact repo source metadata.")
-        if not config.artifact_identifier:
-            raise CloudProviderError("HF Jobs requires an artifact bucket identifier.")
-
-        timestamp = timestamp or self._new_run_timestamp()
-        artifact_prefix = self._build_artifact_prefix(config, timestamp)
-
-        # Read project-specific deps from cloud_config.yaml (single source of truth)
-        cloud_config_path = self.repo_root / "Trainers" / "cloud" / "cloud_config.yaml"
-        project_deps = load_project_deps(cloud_config_path)
-        checkout_steps = build_repo_checkout_steps(
-            RepoCheckoutSpec(
-                url=config.repo_url,
-                branch=config.repo_branch,
-                commit=config.repo_commit,
-            )
-        )
-
-        python_cmd = "$(command -v python3 || command -v python)"
-        parts = [
-            # Install project-specific deps only; unsloth, trl, transformers,
-            # datasets, peft, and PyTorch are pre-installed in the Docker image
-            f"{python_cmd} -m pip install --upgrade {' '.join(project_deps)}",
-            "mkdir -p /tmp/hf-bucket-sync-site",
-            f"{python_cmd} -m pip install --upgrade --target /tmp/hf-bucket-sync-site huggingface_hub>=1.5.0 hf_transfer",
-            f"export HF_BUCKET_SYNC_PYTHON={python_cmd}",
-            "export HF_BUCKET_SYNC_PYTHONPATH=/tmp/hf-bucket-sync-site",
-            f"export CLOUD_PROVIDER={config.provider}",
-            f"export CLOUD_GPU_TYPE={config.hf_flavor or config.gpu_type}",
-            # Enable fast HF transfers
-            "export HF_HUB_ENABLE_HF_TRANSFER=1",
-            *checkout_steps,
-        ]
-        if config.pip_packages:
-            quoted_pip_packages = " ".join(shlex.quote(pkg) for pkg in config.pip_packages)
-            parts.append(f"{python_cmd} -m pip install --upgrade {quoted_pip_packages}")
-
-        training_args: list[str] = []
-        if config.model_name:
-            training_args.extend(["--model-name", config.model_name])
-        if config.dataset_name:
-            training_args.extend(["--dataset-name", config.dataset_name])
-        if config.dataset_file:
-            dataset_file_arg = config.dataset_file
-            if config.dataset_name and dataset_file_arg.startswith(f"{config.dataset_name}/"):
-                dataset_file_arg = dataset_file_arg[len(config.dataset_name) + 1 :]
-            training_args.extend(["--dataset-file", dataset_file_arg])
-        if config.batch_size is not None:
-            training_args.extend(["--batch-size", str(config.batch_size)])
-        if config.gradient_accumulation_steps is not None:
-            training_args.extend(["--gradient-accumulation", str(config.gradient_accumulation_steps)])
-        if config.learning_rate:
-            training_args.extend(["--learning-rate", str(config.learning_rate)])
-        if config.epochs is not None:
-            training_args.extend(["--num-epochs", str(config.epochs)])
-        if config.max_steps is not None:
-            training_args.extend(["--max-steps", str(config.max_steps)])
-        if config.max_seq_length is not None:
-            training_args.extend(["--max-seq-length", str(config.max_seq_length)])
-        if config.method == "sft" and config.load_in_4bit is not None:
-            training_args.append("--load-in-4bit" if config.load_in_4bit else "--no-load-in-4bit")
-        if config.method == "sft" and config.lora_r is not None:
-            training_args.extend(["--lora-r", str(config.lora_r)])
-        if config.method == "sft" and config.lora_alpha is not None:
-            training_args.extend(["--lora-alpha", str(config.lora_alpha)])
-        if config.method == "sft" and config.lora_dropout is not None:
-            training_args.extend(["--lora-dropout", str(config.lora_dropout)])
-        if config.method == "sft" and config.use_dora:
-            training_args.append("--use-dora")
-        if config.method == "sft" and config.use_rslora:
-            training_args.append("--use-rslora")
-        if config.method == "sft" and config.init_lora_weights is not None:
-            training_args.extend(["--init-lora-weights", str(config.init_lora_weights)])
-        if config.method == "sft" and config.lora_target_modules:
-            target_modules_arg = (
-                config.lora_target_modules
-                if isinstance(config.lora_target_modules, str)
-                else ",".join(config.lora_target_modules)
-            )
-            training_args.extend(["--lora-target-modules", target_modules_arg])
-        if config.method == "sft" and config.evolutionary_enabled:
-            training_args.append("--evolutionary-enabled")
-        if config.method == "sft" and config.evolutionary_candidates is not None:
-            training_args.extend(["--evolutionary-candidates", str(config.evolutionary_candidates)])
-        if config.method == "sft" and config.evolutionary_eval_batch_size is not None:
-            training_args.extend(["--evolutionary-eval-batch-size", str(config.evolutionary_eval_batch_size)])
-        if config.method == "sft" and config.evolutionary_validation_config:
-            training_args.extend(["--evolutionary-validation-config", str(config.evolutionary_validation_config)])
-        if config.method == "sft" and config.evolutionary_strategy:
-            training_args.extend(["--evolutionary-strategy", str(config.evolutionary_strategy)])
-        if config.method == "sft" and config.evolutionary_noise_scale is not None:
-            training_args.extend(["--evolutionary-noise-scale", str(config.evolutionary_noise_scale)])
-        if config.method == "sft" and config.evolutionary_max_grad_norm is not None:
-            training_args.extend(["--evolutionary-max-grad-norm", str(config.evolutionary_max_grad_norm)])
-        if config.method == "sft" and config.evolutionary_scale_factors:
-            training_args.extend(
-                ["--evolutionary-scale-factors", ",".join(str(value) for value in config.evolutionary_scale_factors)]
-            )
-        if config.method == "sft" and config.evolutionary_selection_method:
-            training_args.extend(["--evolutionary-selection-method", str(config.evolutionary_selection_method)])
-        if config.method == "sft" and config.evolutionary_min_improvement is not None:
-            training_args.extend(["--evolutionary-min-improvement", str(config.evolutionary_min_improvement)])
-        if config.method == "sft" and config.evolutionary_min_relative_improvement is not None:
-            training_args.extend(["--evolutionary-min-relative-improvement", str(config.evolutionary_min_relative_improvement)])
-        if config.method == "sft" and config.evolutionary_noise_floor_epsilon is not None:
-            training_args.extend(["--evolutionary-noise-floor-epsilon", str(config.evolutionary_noise_floor_epsilon)])
-        if config.method == "sft" and config.evolutionary_eval_frequency is not None:
-            training_args.extend(["--evolutionary-eval-frequency", str(config.evolutionary_eval_frequency)])
-        if config.method == "sft" and config.evolutionary_warmup_steps is not None:
-            training_args.extend(["--evolutionary-warmup-steps", str(config.evolutionary_warmup_steps)])
-        if config.method == "sft" and config.evolutionary_cache_baseline is not None:
-            training_args.append("--evolutionary-cache-baseline" if config.evolutionary_cache_baseline else "--evolutionary-no-cache-baseline")
-        if config.method == "sft" and config.evolutionary_log_candidates is not None:
-            training_args.append("--evolutionary-log-candidates" if config.evolutionary_log_candidates else "--evolutionary-no-log-candidates")
-        if config.method == "sft" and config.evolutionary_log_selected is not None:
-            training_args.append("--evolutionary-log-selected" if config.evolutionary_log_selected else "--evolutionary-no-log-selected")
-        training_args_str = ""
-        if config.method == "grpo":
-            output_dir = f"/workspace/repo/Trainers/{get_canonical_trainer_dir_name(config.method)}/env_grpo_output/{timestamp}"
-            training_args.extend(["--output-dir", output_dir])
-        if training_args:
-            training_args_str = " " + " ".join(shlex.quote(arg) for arg in training_args)
-
-        if config.method == "grpo":
-            parts.extend(
-                self._build_env_grpo_steps(
-                    config=config,
-                    timestamp=timestamp,
-                    artifact_prefix=artifact_prefix,
-                    python_cmd=python_cmd,
-                    training_args=training_args_str,
-                )
-            )
-        else:
-            parts.append(
-                f"cd /workspace/repo/Trainers/{get_canonical_trainer_dir_name(config.method)} && "
-                "python "
-                f"{self._script_name_for_method(config.method)} "
-                f"--run-timestamp {timestamp} "
-                f"--output-root {config.artifact_mount_path} "
-                f"--cloud-provider {config.provider} "
-                f"--artifact-backend {config.artifact_backend} "
-                f"--artifact-bucket {config.artifact_identifier} "
-                f"--artifact-prefix {artifact_prefix} "
-                f"{'--publish-final-model' if config.publish_final_model else ''} "
-                f"{f'--publish-target-repo {config.publish_target_repo}' if config.publish_target_repo else ''}"
-                f"{training_args_str}"
-            )
-
-        return " && ".join(parts)
-
-    def _build_env_grpo_steps(
-        self,
-        *,
-        config: CloudTrainingConfig,
-        timestamp: str,
-        artifact_prefix: str,
-        python_cmd: str,
-        training_args: str,
-    ) -> list[str]:
-        with open(config.config_path, encoding="utf-8") as f:
-            raw_config = yaml.safe_load(f) or {}
-
-        runtime_cfg = ((raw_config.get("env_training") or {}).get("runtime") or {})
-        venv_dir = str(runtime_cfg.get("isolated_venv_dir") or "/workspace/.venvs/grpo-openenv")
-        python_packages = list(runtime_cfg.get("python_packages") or [])
-        project_pip_deps = list(runtime_cfg.get("project_pip_deps") or [])
-        install_args = " ".join(shlex.quote(str(part)) for part in ["pip", "setuptools", "wheel", *project_pip_deps, *python_packages])
-        trainer_dir = f"/workspace/repo/Trainers/{get_canonical_trainer_dir_name(config.method)}"
-        output_dir = f"{trainer_dir}/env_grpo_output/{timestamp}"
-        remote_uri = self._build_remote_run_uri(config, artifact_prefix)
-
-        return [
-            "mkdir -p /tmp/grpo-openenv-bootstrap",
-            f"{python_cmd} -m pip install --upgrade --target /tmp/grpo-openenv-bootstrap virtualenv",
-            "export PYTHONPATH=/tmp/grpo-openenv-bootstrap:$PYTHONPATH",
-            f"{python_cmd} -m virtualenv --no-download {shlex.quote(venv_dir)}",
-            f". {shlex.quote(venv_dir)}/bin/activate",
-            f"python -m pip install --upgrade {install_args}",
-            f"cd {trainer_dir} && python {self.ENV_GRPO_SCRIPT_NAME} --config {trainer_dir}/configs/{self.ENV_GRPO_CONFIG_NAME}{training_args}",
-            (
-                f"if [ -d {shlex.quote(output_dir)} ]; then "
-                f"cd /workspace/repo && "
-                f"PYTHONPATH=/tmp/hf-bucket-sync-site:$PYTHONPATH python -m shared.hf_bucket_sync_helper "
-                f"{shlex.quote(output_dir)} {shlex.quote(remote_uri)}; "
-                "fi"
-            ),
-        ]
 
 
 def _parse_timeout(timeout_str: str) -> float:
