+++ tuner/handlers/stages/__init__.py
@@ -1,0 +1,17 @@
+"""Experiment stage runners for the HF Jobs experiment lifecycle.
+
+Located at tuner/handlers/stages/__init__.py.
+Re-exports all stage runner classes for convenient importing.
+Each runner conforms to the StageRunner Protocol defined in
+shared/experiment_tracking and is used by ExperimentOrchestrator.
+"""
+
+from .hf_eval_stage import HFEvalStageRunner
+from .hf_loss_stage import HFLossStageRunner
+from .hf_training_stage import HFTrainingStageRunner
+
+__all__ = [
+    "HFEvalStageRunner",
+    "HFLossStageRunner",
+    "HFTrainingStageRunner",
+]
